Move option coercion helper to plugin utils for shared use
