--- src/plugin/src/printer/option-utils.js
@@ -1,1 +1,0 @@
-export { coercePositiveIntegerOption } from "../options/option-utils.js";
