+++ src/plugin/src/utils/option-utils.js
